agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pysnmp==4.4.12
pyasn1==0.4.8
python-dotenv==1.0.0
requests==2.31.0
